Implemented Player Spawn Points
Changed World to Resource

+++ src/constants.rs
@@ -7,6 +7,6 @@ pub const MARGINS: f32 = 0.00;
 /// (1000 / rows) - margins; // Size of a tile, all tiles are square
 pub const TILE_SIZE: f32 = 1.0;
 /// Texture size in pixels
-pub const TEXTURE_SIZE: f32 = 32.0;
+pub const TEXTURE_SIZE: usize = 32;
 /// The speed of the player movement
 pub const PLAYER_SPEED: f32 = 4.0;
